backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# OAuth & HTTP
httpx==0.26.0
authlib==1.3.0

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Document Processing
python-docx>=1.1.1
docxtpl==0.17.0
pypandoc==1.12
PyPDF2==3.0.1
pdfplumber==0.10.3
reportlab==4.0.9

# AI/LLM
openai==1.51.0
anthropic==0.39.0

# NLP for skill extraction
spacy==3.7.2
# Download model with: python -m spacy download en_core_web_sm

# Utils
aiofiles==23.2.1
